21/7

+++ wdio.hooks.conf.js
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import deepmerge from 'deepmerge';
-import {config as baseConfig} from '../../../../wdio.conf.js';
+import {config as baseConfig} from './wdio.conf.js';
 
 const passedDirectory = 'screenshots/passed';
 const failedDirectory = 'screenshots/failed';
@@ -12,16 +12,20 @@ function createIfNotExists(dir) {
 }
 
 function deleteFiles(dir) {
-    fs.rm(dir, { recursive: true }, err => {
-        if (err) console.log(err);
-    });
+    if (fs.existsSync(dir)) {
+        fs.rm(dir, { recursive: true }, err => {
+            if (err) console.log(err);
+        });
+    }
 }
 
 export const config = deepmerge(baseConfig, {
     specs: [
-        '../lesson-05*/*.e2e.js'
+        // '../lesson-05*/*.e2e.js'
+        './test/specs/homework/homework.e2e.js'
     ],
-
+    exclude: [ './test/specs/exercise.e2e.js'],
+     
     /*
     Definice potřebných hooků
     */
@@ -33,8 +37,17 @@ export const config = deepmerge(baseConfig, {
         const screenshotName = (`${test.parent}__${test.title}.png`).replace(/ /g, '_');
         if (passed === false) {
             createIfNotExists(failedDirectory);
+            // console.log(`${failedDirectory}/${screenshotName}`);
             browser.saveScreenshot(`${failedDirectory}/${screenshotName}`);
+        
+        }
+        else {
+            createIfNotExists(passedDirectory);
+            // console.log(`${failedDirectory}/${screenshotName}`);
+            browser.saveScreenshot(`${passedDirectory}/${screenshotName}`); 
         }
     }
 
 }, { clone: false })
+
+// npx wdio run wdio.hooks.conf.js
